Finish the Markdown Preview challenge
Although I was encouraged to use a framework like React to do this I
wanted to try and do it without one. I'm relatively familiar with how
React and Redux work but I feel less familiar with native browser API
features, so I thought this was a good opportunity to experiment. All
tests are passing.