Set up structure and state
The drum machine now has the concept of clips, banks and pads
and the main component stores the state of the bank of buttons,
ie. which buttons are associated with which clips and whether
the button is active or not. This may or may not be used in the
actual implementation. It's not yet clear whether the machine
will be used for sequencing or whether it just needs to play a sound
when you click a particular button.