Implement the drum pad structure
The initial layout of the drum pad bank is laid out with some rough
styling to show the eventual intent. This is still a work in progress,
and the audio clips and ability to play them is yet to be implemented
(it should come in the next commit). There was some difficulty in
getting the drum-bad bank to align vertically on the page, which was
impossible with flex-box, so positioning and negative margins had to be
used for that very first container. The rest seems to be working well
with a combination of flex and grid.

Currently, the pad audio element is wrapped in an anchor, though that
may not be ideal depending on how that interacts with the play action.
It's also currently not clear how we trigger playing of the audio clip
using an event on the anchor within the React component.