Add a track name display
Also did some cleaning up of the mark-up based on some (mental) changes
to the planned layout. The `setTimeout` based function to clear the
lable after a certain amount of time is a bit buggy because of the race
condition with the user clicking new buttons faster than the display
disappears. There's probably a better way to do this by using a "stack"
of labels that are waiting to be displayed and clear if a new one is
added—or something like that.