Sets up the initial calculator logic
The calculator is broken up into a couple of components including
Buttons, which are going to be rendered based on some information stored
in a BUTTONS constant. It doesn't seem like this should be stored in
state as it shouldn't ever really change, but I'm not sure if this is
the best place/way for it. A number of button properties will be driven
by this constant, such as:

- Button labels
- Button values
- Button actions (add, subtract, etc.)
- Button types (which will determine colour)
- Button position (on the keypad layout)

It may be better to give them unique IDs and do all the positioning in
CSS, so that may be a forthcoming enhancement.