Implement most calculator logic and appearance
This should probably have been split over two commmits but this
implements the basic layout of the calculator using a CSS grid, renders
all the buttons and implements the actual input/output processing.

Numbers and operators are logged into a buffer in the main component
state, which is then converted into a function when the '=' button is
clicked, displaying the result in the `Display` component.

In future, I will try to break work like this into separate commits that
flow logically with the evolution of the application.