Implement core timing logic
The clock can now count down by seconds from a supplied default time and
update the DOM accordingly. There are both start and stop controls.
The clock uses React's useEffect() to trigger a standard setTimeout set
to 1 second, which will decrement the remaining time in state by one
second. This in turn restarts the loop thanks to the state change
causing React to set a new setTimeout. The concept of a "mode" controls
whether React sets a new Timeout or not.