Implement toggling between Session and Break
The clock now switches to a "Break" once the timing "Session" is over
and starts counting down from whatever the break length is defined as.
It changes a (new) label above the time to indicate what "mode" it is
in, which required moving the previous concept of "mode" to "state".
"State" now refers to whether the clock is running, stopped, etc. and
"mode" refers to whether it's in a Session or a Break.

Some minor tweaks to the settings input fields meant that we can now
either type in values or use the increment/decrement controls to adjust
the time settings.