Simplify states and add a 'beep' on completion
The logic for switching states was a bit more complex than it needed to
be, so this has been simplified and the Clock can really only be running
or not running, in session or in a break. This has resulted in
previously cryptic tests failures passing.

There's now also an audio 'beep' played when the timer reaches zero,
using an HTML audio element.