Move handlers into separate functions
This partially addresses Issue #6 by moving the logic for incrementing
and decrementing the session and break lengths into separate functions
nad out of the "big switch". It also moves formatting of the time output
down the chain to be handled by the components that display it, rather
than further up where the actual data lives.